.NET sample to perform a blob upload using SharedKey via Put Blob REST API
This sample code written in .NET helps to perform a blob upload (PUT Blob) operation using the SharedKey via REST API and not using SDK.
Update the Storage account name and the Access Key in the project. Add the container name and the blob name with its contents. 
Then you are good to perform the upload operation via PUT blob REST API