Create AzureRedisPubSubDemoConsole
Redis Pub Sub provides allow us to use the messaging paradigm by using channels to publish messages and for subscribers to listen for notification of the message. Redis Pub Sub allows multiple subscribers to listen to one or more channels, and to only receive messages that are of interest. It decouples the subscriber from the publisher since the subscriber has no knowledge of whom the publishers are and vice versa there could be multiple publisher not knowing whom the subscribers are.

There are definitely certain restrictions of using Redis Pub Sub as a Messaging System, it will not be like RabbitMQ, Kafka or Azure MessageBus etc. Those message bus are able to store the message for durability or even replay of an old message for consumption. Redis uses a listener model where there are no listeners (subscribers) it will not receive those messages. 


Redis Pub Sub – Operations
PUBLISH channels message: Posts a message to the given channel O(N+M)
SUBSCRIBE [channel]: Subscribe to a given channel for message, O(N) where N is the number of channels to subscribe to
PSUBSCRIBE [channel]: Subscribes the client to the given patterns, O(N) where N is the number of patterns the client is already subscribed to.
PUBSUB CHANNELS pattern: Currently active channels, O(N)
PUBSUB NUMSUB channel: Number of subscribers to the channels provided, O(N)
PUBSUB NUMPAT: Number of subscriptions to all the patterns O(N)
PUNSUBSCRIBE: Unsubscribes the client from a pattern, O(N+M)
UNSUBSCRIBE: Unsubscribes the client from a channel, O(N) where N is the number of clients already subscribed to a channel.